move getConfig to /config
