+++ bin/underreact.js
@@ -2,17 +2,12 @@
 'use strict';
 
 const yargs = require('yargs');
-const path = require('path');
-const chalk = require('chalk');
-const webpack = require('webpack');
 
 const start = require('../commands/start');
 const build = require('../commands/build');
 const serveStatic = require('../commands/serve-static');
 const logger = require('../lib/logger');
-const { DEFAULT_CONFIG_NAME } = require('../lib/constants');
-const createConfig = require('../lib/config');
-const getDefaultConfig = require('../lib/default-underreact.config');
+const config = require('../lib/config');
 
 const configOption = [
   'config',
@@ -80,7 +75,7 @@ function defineStart(y) {
 }
 
 function runStart(argv) {
-  getConfig('start', argv)
+  config('start', argv)
     .then(urc => start(urc))
     .catch(errorOut);
 }
@@ -98,7 +93,7 @@ function defineBuild(y) {
 }
 
 function runBuild(argv) {
-  getConfig('build', argv)
+  config('build', argv)
     .then(urc => build(urc))
     .catch(errorOut);
 }
@@ -111,73 +106,9 @@ function defineServeStatic(y) {
 }
 
 function runServeStatic(argv) {
-  getConfig('serve-static', argv)
+  config('serve-static', argv)
     .then(urc => serveStatic(urc))
     .catch(errorOut);
-}
-
-function getConfig(command, argv) {
-  const configIsNotSpecified = argv.config === undefined;
-  let configPath;
-
-  if (configIsNotSpecified) {
-    configPath = path.join(process.cwd(), DEFAULT_CONFIG_NAME);
-  } else {
-    configPath = path.resolve(argv.config);
-  }
-
-  return userConfigReader(configPath)
-    .catch(() => {
-      // If the user didn't use --config, it's fine if the underreact.config.js
-      // does not exist at the default path.
-      if (configIsNotSpecified) {
-        return () => {};
-      }
-      // But if they did specify a path, tell them it didn't work.
-      throw new Error(
-        `Failed to load underreact configuration module from ${chalk.underline(
-          configPath
-        )}`
-      );
-    })
-    .then(getUserConfig => {
-      let production = true;
-      if (command === 'build') {
-        production = !argv.debug;
-      } else if (command === 'start') {
-        production = argv.production;
-      } else if (argv.env === 'prod') {
-        production = true;
-      }
-
-      // This object is passed as the argument to the config module, if it's a
-      // function.
-      const configModuleContext = {
-        webpack,
-        command,
-        production
-      };
-
-      const userConfig = getUserConfig(configModuleContext);
-      const defaultConfig = getDefaultConfig(
-        Object.assign({}, configModuleContext, {
-          argv,
-          configDir: path.dirname(configPath),
-          stats: argv.stats,
-          port: argv.port
-        })
-      );
-      return createConfig(userConfig, defaultConfig);
-    });
-}
-
-function userConfigReader(configPath) {
-  return Promise.resolve().then(() => {
-    const configModule = require(configPath);
-    return typeof configModule === 'function'
-      ? configModule
-      : () => configModule;
-  });
 }
 
 function errorOut(error) {
