+++ lib/config/createUrc.test.js
@@ -3,7 +3,7 @@
 const path = require('path');
 
 const getDefaultConfig = require('../default-underreact.config');
-const createConfig = require('./createUrc');
+const createUrc = require('./createUrc');
 
 test('normalizes empty configuration', () => {
   const defaultConfig = getDefaultConfig({
@@ -12,7 +12,7 @@ test('normalizes empty configuration', () => {
     port: undefined,
     configDir: path.join(process.cwd(), 'underreact.config.js')
   });
-  expect(createConfig({}, defaultConfig)).toMatchSnapshot();
+  expect(createUrc({}, defaultConfig)).toMatchSnapshot();
 });
 
 test('normalizes fancy configuration', () => {
@@ -23,7 +23,7 @@ test('normalizes fancy configuration', () => {
     configDir: path.join(process.cwd(), 'underreact.config.js')
   });
   expect(
-    createConfig(
+    createUrc(
       {
         polyfills: {
           promise: true,
